requirements.txt
Django==5.0.3
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.3.0
django-ninja==1.1.0
pydantic==2.6.4
SQLAlchemy==2.0.29
psycopg2==2.9.9

psycopg==3.1.18

redis==5.0.3
celery==5.3.6

flake8==7.0.0
isort==5.13.2
black==24.2.0

nats-py==2.7.2
pytube==15.0.0
protobuf==3.20.0

langchain==0.1.16
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-community==0.0.34
langchainhub==0.1.15 
langchain-chroma==0.1.0